chore: restructured project files and folders
