+++ src/App.js
@@ -1,8 +1,8 @@
 import React from 'react';
 import './App.scss';
-import LaunchCard  from './components/Launches/LaunchCard';
+import LaunchCard  from './components/Launches/LaunchCard/LaunchCard.jsx';
 import NavBar from './components/navbar/NavBar';
-import PastLaunches from './components/Launches/PastLaunches';
+import PastLaunches from './components/Launches/pastLaunches/PastLaunches';
 
 
 function App() {
