+++ src/components/navbar/NavBar.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-
+import './styles.scss'
 function NavBar() {
 
     return (
